Split Announce away from ExecutionParameters
